5251 - Split out the processing to new file
